+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/npc/NPCSpawner.java
@@ -4,7 +4,9 @@ import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
+import org.jspecify.annotations.NonNull;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -17,7 +19,7 @@ public class NPCSpawner {
         this.commandBuilder = new NPCMannequinCommandBuilder();
     }
 
-    public Optional<UUID> spawnMannequin(NPC npc) {
+    public Optional<UUID> spawnMannequin(@NonNull NPC npc) {
         World world = Bukkit.getWorld(npc.getWorldName());
         if (world == null) {
             return Optional.empty();
@@ -33,7 +35,7 @@ public class NPCSpawner {
 
         Bukkit.dispatchCommand(playerInWorld.get(), summonCommand);
 
-        return findNewlySpawnedMannequin(world, spawnLocation);
+        return findNewlySpawnedMannequin(world, spawnLocation).map(this::applyMannequinAttributes);
     }
 
     public void despawnMannequin(UUID entityUUID) {
@@ -43,7 +45,17 @@ public class NPCSpawner {
         }
     }
 
-    private Optional<UUID> findNewlySpawnedMannequin(World world, Location spawnLocation) {
+    private UUID applyMannequinAttributes(UUID entityUUID) {
+        Entity entity = Bukkit.getEntity(entityUUID);
+
+        if (entity instanceof LivingEntity livingEntity) {
+            livingEntity.setCanPickupItems(false);
+            livingEntity.setCollidable(false);
+        }
+        return entityUUID;
+    }
+
+    private @NonNull Optional<UUID> findNewlySpawnedMannequin(@NonNull World world, Location spawnLocation) {
         return world.getEntities().stream()
                 .filter(entity -> entity.getType().name().equalsIgnoreCase("MANNEQUIN"))
                 .filter(entity -> isWithinSpawnRadius(entity.getLocation(), spawnLocation))
@@ -51,7 +63,7 @@ public class NPCSpawner {
                 .findFirst();
     }
 
-    private boolean isWithinSpawnRadius(Location entityLocation, Location targetLocation) {
+    private boolean isWithinSpawnRadius(@NonNull Location entityLocation, Location targetLocation) {
         double spawnRadiusTolerance = 0.5;
         return entityLocation.distanceSquared(targetLocation) <= spawnRadiusTolerance;
     }
